+++ src/plugins/data_source_management/public/auth_registry/authentication_methods_registry.ts
@@ -8,8 +8,9 @@ import { EuiSuperSelectOption } from '@elastic/eui';
 
 export interface AuthenticationMethod {
   name: string;
-  credentialForm: React.JSX.Element;
   credentialSourceOption: EuiSuperSelectOption<string>;
+  credentialForm?: React.JSX.Element;
+  crendentialFormField?: { [key: string]: string };
 }
 
 export type IAuthenticationMethodRegistery = Omit<